database cleanups next iteration

+++ backend/app/alembic/versions/043_create_global_schema_and_languages_table.py
@@ -25,7 +25,7 @@ def upgrade():
         "languages",
         sa.Column(
             "id",
-            sa.BigInteger(),
+            sa.Integer(),
             sa.Identity(always=False),
             primary_key=True,
             comment="Unique identifier for the language",
